updating for lid detection test

+++ Block_wise_GPS_correction.py
@@ -4266,7 +4266,7 @@ class Field:
 
 		if settings.use_temp_matching:
 			results_sorted = sorted(results, reverse=True, key=lambda tup: tup[6])
-			max_count_lids = max(len(results_sorted)/5,28)
+			max_count_lids = max(int(len(results_sorted)/5),28)
 			results = [(x,y,r,l,pn,crd) for x,y,r,l,pn,crd,score in results_sorted[:max_count_lids]]
 
 		for x,y,r,l,pn,crd,score in results:
